Crate cosmwasm_schema

Re-exports

pub use schemars;

pub use serde;

Macros

generate_api

Generates an Api for the contract. The body describes the message types exported in the schema and allows setting contract name and version overrides.

schema_for

Generates a RootSchema for the given type using default settings.

write_api

Takes care of generating the interface description file for a contract. The body describes the message types included and allows setting contract name and version overrides.

Structs

Api

Rust representation of a contract’s API.

Enums

IntegrityError

Constants

IDL_VERSION

The version of the CosmWasm IDL.

Traits

QueryResponses

A trait for tying QueryMsg variants (different contract queries) to their response types. This is mostly useful for the generated contracted API description when using cargo schema.

Functions

combine_subqueries

Combines multiple response schemas into one. Panics if there are name collisions. Used internally in the implementation of QueryResponses when using #[query_responses(nested)]

export_schema

export_schema_with_title

remove_schemas

Attribute Macros

cw_serde

An attribute macro that annotates types with things they need to be properly (de)serialized for use in CosmWasm contract messages and/or responses, and also for schema generation.

Derive Macros

QueryResponses